Move handle_debugger to debugger module Imolemented as an external impl for gameboy since it needs all its state
