+++ core/src/cpu/mod.rs
@@ -3,5 +3,3 @@ pub mod register;
 pub mod opcodes;
 pub mod flag;
 pub mod opcode_runner;
-#[cfg(feature = "dbg")]
-pub mod disassembler;
