+++ core/src/lib.rs
@@ -11,4 +11,10 @@ pub mod utils;
 #[cfg(feature = "dbg")]
 pub mod debugger;
 
-pub use utils::GB_FREQUENCY;
+pub use {
+    machine::gameboy::GameBoy,
+    ppu::gfx_device::GfxDevice,
+    apu::audio_device::AudioDevice,
+    keypad::joypad_provider::JoypadProvider,
+    utils::GB_FREQUENCY, 
+};
