+++ core/src/mmu/mod.rs
@@ -13,3 +13,9 @@ pub mod vram_dma_controller;
 
 pub use external_memory_bus::GB_BOOT_ROM_SIZE;
 pub use external_memory_bus::GBC_BOOT_ROM_SIZE;
+
+pub trait Memory{
+    fn read(&mut self, address:u16, m_cycles:u8)->u8;
+    fn write(&mut self, address:u16, value:u8, m_cycles:u8);
+    fn set_double_speed_mode(&mut self, state:bool);
+}
